pyproject.toml

[project]
name = "goal_rpg"
version = "0.1.0"
description = ""
authors = [
    {name = "Matt Zhang", email = "set.stun@gmail.com"},
]
dependencies = [
    "py-gpt-interface>=0.1.0",
]
requires-python = ">=3.11"
readme = "README.md"
license = {text = "MIT"}

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

